Move is_legal to engine/check.py
